smart_parking/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
daphne==4.2.1
distro==1.9.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
easyocr==1.7.2
exceptiongroup==1.3.0
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
openai==1.102.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyserial==3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
redis==6.4.0
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
service-identity==24.2.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.5.10
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.182
ultralytics-thop==2.0.16
uritemplate==4.2.0
urllib3==2.5.0
zope.interface==7.2
